+++ src/main/java/com/example/practice/auth/controller/dto/LoginRequest.java
@@ -1,0 +1,5 @@
+package com.example.practice.auth.controller.dto;
+
+public record LoginRequest(String loginId, String password) {
+
+}
