+++ src/main/java/com/example/practice/auth/controller/dto/LoginResponse.java
@@ -1,0 +1,9 @@
+package com.example.practice.auth.controller.dto;
+
+public record LoginResponse(String accessToken, String tokenType) {
+
+
+    public static LoginResponse withUUID(String uuid) {
+        return new LoginResponse(uuid, "Bearer");
+    }
+}
