+++ src/main/java/com/example/practice/auth/util/AuthTokenUtils.java
@@ -1,0 +1,21 @@
+package com.example.practice.auth.util;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class AuthTokenUtils { // AuthUtil AuthTokenUtil
+    private static final String PREFIX = "Bearer ";
+
+    // parsing(accessToken(Bearer + SessionKey) -> SessionKey)
+
+    public static boolean isValidToken(String accessToken) {
+        return accessToken.startsWith(PREFIX);
+    }
+    public static String getSessionKey(String accessToken) {
+        return accessToken.substring(PREFIX.length()).trim();
+    }
+    //
+
+
+}
