--- src/main/java/com/example/practice/controller/HelloController.java
@@ -1,23 +1,0 @@
-package com.example.practice.controller;
-
-import com.example.practice.service.HelloService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HelloController {
-
-    private final HelloService helloService;
-
-    @Autowired
-    public HelloController(HelloService helloService) {
-        this.helloService = helloService;
-    }
-
-    @GetMapping("/hello")
-    public String hello() {
-        return helloService.getString();
-    }
-
-}
