--- src/main/java/com/example/practice/controller/ParamController.java
@@ -1,4 +1,0 @@
-package com.example.practice.controller;
-
-public class ParamController {
-}
