--- src/main/java/com/example/practice/controller/dto/CreatePostRequest.java
@@ -1,16 +1,0 @@
-package com.example.practice.controller.dto;
-
-public class CreatePostRequest {
-    public String title;
-    public String content;
-    public String author;
-
-    @Override
-    public String toString() {
-        return "CreatePostRequest{" +
-                "title='" + title + '\'' +
-                ", content='" + content + '\'' +
-                ", Author='" + author + '\'' +
-                '}';
-    }
-}
