--- src/main/java/com/example/practice/controller/dto/LoginRequest.java
@@ -1,5 +1,0 @@
-package com.example.practice.controller.dto;
-
-public record LoginRequest(String loginId, String password) {
-
-}
