--- src/main/java/com/example/practice/controller/dto/LoginResponse.java
@@ -1,9 +1,0 @@
-package com.example.practice.controller.dto;
-
-public record LoginResponse(String accessToken, String tokenType) {
-
-
-    public static LoginResponse withUUID(String uuid) {
-        return new LoginResponse(uuid, "Bearer");
-    }
-}
