--- src/main/java/com/example/practice/controller/dto/MemberRequest.java
@@ -1,4 +1,0 @@
-package com.example.practice.controller.dto;
-
-public record MemberRequest(String loginId, String password, String name) {
-}
