--- src/main/java/com/example/practice/controller/dto/PostRequest.java
@@ -1,3 +1,0 @@
-package com.example.practice.controller.dto;
-public record PostRequest(String title, String content, String author) {
-}
