--- src/main/java/com/example/practice/global/exception/ErrorResponse.java
@@ -1,11 +1,0 @@
-package com.example.practice.global.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ErrorResponse {
-    private int status;
-    private String message;
-}
