--- src/main/java/com/example/practice/global/exception/NotFoundException.java
@@ -1,7 +1,0 @@
-package com.example.practice.global.exception;
-
-public class NotFoundException extends RuntimeException {
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
