--- src/main/java/com/example/practice/global/exception/UnauthorizedException.java
@@ -1,7 +1,0 @@
-package com.example.practice.global.exception;
-
-public class UnauthorizedException extends RuntimeException {
-    public UnauthorizedException(String message) {
-        super(message);
-    }
-}
