+++ src/main/java/com/example/practice/hello/service/HelloService.java
@@ -1,0 +1,11 @@
+package com.example.practice.hello.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class HelloService {
+
+    public String getString() {
+        return "hi String";
+    }
+}
