+++ src/main/java/com/example/practice/member/controller/dto/MemberRequest.java
@@ -1,0 +1,4 @@
+package com.example.practice.member.controller.dto;
+
+public record MemberRequest(String loginId, String password, String name) {
+}
