+++ src/main/java/com/example/practice/member/controller/dto/MemberResponse.java
@@ -1,0 +1,9 @@
+package com.example.practice.member.controller.dto;
+
+import com.example.practice.member.entity.MemberEntity;
+
+public record MemberResponse(Long id, String loginId, String name) {
+    public static MemberResponse from(MemberEntity saveMember) {
+        return new MemberResponse(saveMember.getId(), saveMember.getLoginId(), saveMember.getName());
+    }
+}
