--- src/main/java/com/example/practice/service/HelloService.java
@@ -1,11 +1,0 @@
-package com.example.practice.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class HelloService {
-
-    public String getString() {
-        return "hi String";
-    }
-}
